% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApplyGSVA.R
\name{ApplyGSVA}
\alias{ApplyGSVA}
\title{Apply GSVA Method to a Expression Matrix List}
\usage{
ApplyGSVA(
  GeneSets,
  group_col,
  gene_col,
  ExprMatList,
  method = c("ssgsea", "gsva", "zscore", "plage"),
  kcdf = c("Gaussian", "Poisson"),
  ...
)
}
\arguments{
\item{GeneSets}{a \code{data.frame} containing groups and genes.}

\item{group_col}{the column name for groups.}

\item{gene_col}{the column name for genes.}

\item{ExprMatList}{a \code{list} of expression matrix represented with \code{tibble}.
See also \code{\link[=GenTibbleList]{GenTibbleList()}}.}

\item{method}{Method to employ in the estimation of gene-set enrichment scores per sample. By default
                this is set to \code{gsva} (\enc{H盲nzelmann}{Hanzelmann} et al, 2013) and other options are
                \code{ssgsea} (Barbie et al, 2009), \code{zscore} (Lee et al, 2008) or \code{plage}
                (Tomfohr et al, 2005). The latter two standardize first expression profiles into z-scores
                over the samples and, in the case of \code{zscore}, it combines them together as their sum
                divided by the square-root of the size of the gene set,
                while in the case of \code{plage} they are used to calculate the singular value decomposition
                (SVD) over the genes in the gene set and use the coefficients of the first right-singular vector
                as pathway activity profile.}

\item{kcdf}{Character string denoting the kernel to use during the non-parametric estimation of the
              cumulative distribution function of expression levels across samples when \code{method="gsva"}.
              By default, \code{kcdf="Gaussian"} which is suitable when input expression values are continuous,
              such as microarray fluorescent units in logarithmic scale, RNA-seq log-CPMs, log-RPKMs or log-TPMs.
              When input expression values are integer counts, such as those derived from RNA-seq experiments,
              then this argument should be set to \code{kcdf="Poisson"}.}

\item{...}{other arguments passing to \code{\link[GSVA:gsva]{GSVA::gsva()}}}
}
\value{
a result \code{list}
}
\description{
Apply GSVA Method to a Expression Matrix List
}
\examples{
data("example_gsets")
data("example_data")

ExprList <- GenTibbleList(list(example_data))
res <- ApplyGSVA(example_gsets,
  group_col = "Cell_type",
  gene_col = "Symbol", ExprMatList = ExprList, method = "gsva"
)
}
