% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenTibbleList.R
\name{GenTibbleList}
\alias{GenTibbleList}
\title{Generate a List of Tibbles from a List of Gene Expression Sets}
\usage{
GenTibbleList(ExpressionSetList)
}
\arguments{
\item{ExpressionSetList}{a list of \code{\link[Biobase:ExpressionSet]{Biobase::ExpressionSet()}}}
}
\value{
a \code{list}
}
\description{
Generate a List of Tibbles from a List of Gene Expression Sets
}
\examples{
data("example_data")
GenTibbleList(list(example_data))
}
\author{
Shixiang Wang
}
